+++ src/wp/ap/gtk/ap_UnixDialog_Document.cpp
@@ -1,5 +1,4 @@
 /* Abinova
- * Copyright (C) 2026 AbiSource, Inc.
  * Copyright (C) 2025-2026 Abinova contributors
  *
  * This program is free software; you can redistribute it and/or