+++ src/wp/ap/gtk/ap_UnixRibbon.cpp
@@ -1,6 +1,5 @@
 /* -*- mode: C++; tab-width: 4; c-basic-offset: 4; indent-tabs-mode:t -*- */
 /* Abinova
- * Copyright (C) 2026 AbiSource, Inc.
  * Copyright (C) 2025-2026 Abinova contributors
  *
  * This program is free software; you can redistribute it and/or